@@ -1,73 +1,103 @@
 {
-  "name": "MERNjs",
-  "version": "1.0.0",
-  "description": "A Open-Source Full-Stack Solution For MERN (MongoDB, ExpressJS, ReactJS and NodeJS) Applications.",
-  "repository": {
-    "type": "git",
-    "url": "git+ssh://git@github.com:IamMohaiminul/MERNjs.git"
-  },
-  "author": "Muhammad Mohaiminul Islam <IamMohaiminul>",
-  "license": "MIT",
+  "name": "mernjs",
+  "version": "2.0.0",
+  "description": "An open-source full-stack JavaScript solution For MERN (MongoDB, ExpressJS, ReactJS and NodeJS) application",
+  "keywords": [
+    "NodeJS",
+    "ExpressJS",
+    "ReactJS",
+    "MongoDB",
+    "ReduxJS",
+    "Webpack",
+    "babel"
+  ],
+  "author": "Muhd Mohaiminul Islam <IamMohaiminul>",
   "engines": {
-    "node": "~6.x"
+    "node": "12.x"
   },
   "scripts": {
     "start": "node ./build",
     "debug": "node --inspect ./build",
-    "build": "npm run clean:server && npm run build:server && npm run build:client",
-    "clean:server": "rm -rf build && mkdir build",
+    "build:client": "webpack",
     "build:server": "babel -d ./build ./server -s",
-    "build:client": "webpack --progress --profile",
-    "live:server": "nodemon --exec babel-node ./server",
-    "live:server:debug": "DEBUG=MERNjs:* npm run live:server",
-    "live:client": "webpack --watch --progress --profile"
+    "build": "rimraf ./build && concurrently \"npm:build:client\" \"npm:build:server\"",
+    "watch:client": "webpack --watch",
+    "watch:server": "DEBUG=MERNjs:* nodemon --exec babel-node ./server",
+    "watch": "concurrently \"npm:watch:client\" \"npm:watch:server\""
   },
   "dependencies": {
-    "app-module-path": "~2.2.0",
-    "axios": "~0.15.3",
-    "bcrypt-nodejs": "0.0.3",
-    "body-parser": "~1.17.1",
-    "config": "~1.25.1",
-    "cookie-parser": "~1.4.3",
-    "cors": "~2.8.1",
-    "debug": "~2.6.3",
-    "express": "~4.15.2",
-    "jsonwebtoken": "~7.3.0",
-    "lodash": "~4.17.4",
-    "moment": "~2.18.1",
-    "mongoose": "~4.9.1",
-    "morgan": "~1.8.1",
-    "pug": "~2.0.0-beta11",
-    "serve-favicon": "~2.4.1"
+    "@babel/polyfill": "^7.10.1",
+    "app-module-path": "^2.2.0",
+    "axios": "^0.19.2",
+    "bcryptjs": "^2.4.3",
+    "body-parser": "^1.19.0",
+    "cookie-parser": "^1.4.5",
+    "cors": "^2.8.5",
+    "debug": "^4.1.1",
+    "dotenv": "^8.2.0",
+    "express": "^4.17.1",
+    "jsonwebtoken": "^8.5.1",
+    "lodash": "^4.17.15",
+    "moment": "^2.26.0",
+    "mongoose": "^5.9.18",
+    "morgan": "^1.10.0",
+    "pug": "^3.0.0",
+    "serve-favicon": "^2.5.0",
+    "toastr": "^2.1.4"
   },
   "devDependencies": {
-    "babel-cli": "~6.24.0",
-    "babel-core": "~6.24.0",
-    "babel-loader": "~6.4.1",
-    "babel-preset-env": "~1.2.2",
-    "babel-preset-react": "~6.23.0",
-    "bootstrap": "~3.3.7",
-    "clean-webpack-plugin": "~0.1.16",
-    "compression-webpack-plugin": "~0.3.2",
-    "css-loader": "~0.27.3",
-    "extract-text-webpack-plugin": "~2.1.0",
-    "file-loader": "~0.10.1",
-    "font-awesome": "~4.7.0",
-    "jquery": "~2.2.4",
-    "node-sass": "~4.5.1",
-    "nodemon": "~1.11.0",
-    "react": "~15.4.2",
-    "react-dom": "~15.4.2",
-    "react-redux": "~5.0.3",
-    "react-router": "~3.0.2",
-    "redux": "~3.6.0",
-    "redux-logger": "~2.10.2",
-    "redux-thunk": "~2.2.0",
-    "sass-loader": "~6.0.3",
-    "style-loader": "~0.15.0",
-    "toastr": "~2.1.2",
-    "url-loader": "~0.5.8",
-    "webpack": "~2.3.0"
+    "@babel/cli": "^7.10.1",
+    "@babel/core": "^7.10.2",
+    "@babel/node": "^7.10.1",
+    "@babel/preset-env": "^7.10.2",
+    "@babel/preset-react": "^7.10.1",
+    "autoprefixer": "^9.8.0",
+    "babel-loader": "^8.1.0",
+    "bootstrap": "^4.5.0",
+    "clean-webpack-plugin": "^3.0.0",
+    "compression-webpack-plugin": "^4.0.0",
+    "concurrently": "^5.2.0",
+    "css-loader": "^3.5.3",
+    "dotenv-webpack": "^1.8.0",
+    "eslint": "^6.8.0",
+    "eslint-config-airbnb": "^18.1.0",
+    "eslint-config-prettier": "^6.11.0",
+    "eslint-plugin-import": "^2.21.2",
+    "eslint-plugin-jsx-a11y": "^6.2.3",
+    "eslint-plugin-react": "^7.20.0",
+    "eslint-plugin-react-hooks": "^2.5.1",
+    "file-loader": "^6.0.0",
+    "font-awesome": "^4.7.0",
+    "history": "^4.10.1",
+    "jquery": "^3.5.1",
+    "mini-css-extract-plugin": "^0.9.0",
+    "node-sass": "^4.14.1",
+    "nodemon": "^2.0.4",
+    "optimize-css-assets-webpack-plugin": "^5.0.3",
+    "path": "^0.12.7",
+    "popper.js": "^1.16.1",
+    "postcss-loader": "^3.0.0",
+    "precss": "^4.0.0",
+    "prop-types": "^15.7.2",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
+    "react-redux": "^7.2.0",
+    "react-router": "^5.2.0",
+    "react-router-dom": "^5.2.0",
+    "redux": "^4.0.5",
+    "redux-logger": "^3.0.6",
+    "redux-thunk": "^2.3.0",
+    "rimraf": "^3.0.2",
+    "sass-loader": "^7.3.1",
+    "terser-webpack-plugin": "^3.0.4",
+    "url-loader": "^4.1.0",
+    "webpack": "^4.43.0",
+    "webpack-cli": "^3.3.11"
   },
+  "repository": {
+    "type": "git",
+    "url": "git+ssh://git@github.com:IamMohaiminul/MERNjs.git"
+  },
+  "license": "MIT",
   "homepage": "http://mern.js.org/"
 }
