@@ -1,87 +1,106 @@
 import React, { Component } from 'react';
-import ReactDOM from 'react-dom';
-import { Link, browserHistory } from 'react-router';
+import PropTypes from 'prop-types';
+import { Link } from 'react-router-dom';
 import toastr from 'toastr';
 
 class LoginComponent extends Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      emailAddress: '',
+      password: '',
+    };
+
+    this.handleInputChange = this.handleInputChange.bind(this);
+    this.handleFormSubmit = this.handleFormSubmit.bind(this);
+  }
+
+  handleInputChange(event) {
+    const { name } = event.target;
+    const value = event.target.type === 'checkbox' ? event.target.checked : event.target.value;
+    this.setState({ [name]: value });
+  }
+
+  handleFormSubmit(event) {
+    event.preventDefault();
+    const { loginUser, history } = this.props;
+    const { emailAddress, password } = this.state;
+
+    if (emailAddress.trim().length < 1) {
+      toastr.warning('Email Address is required.', 'MERNjs');
+      return false;
+    }
+    if (password.trim().length < 1) {
+      toastr.warning('Password is required.', 'MERNjs');
+      return false;
+    }
+
+    loginUser(
+      {
+        emailAddress,
+        password,
+      },
+      (err, res) => {
+        if (err) {
+          toastr.error(err.message, 'MERNjs');
+          return false;
+        }
+        toastr.success(res.message, 'MERNjs');
+        history.push('/admin');
+        return true;
+      },
+    );
+    return null;
+  }
+
   render() {
+    const { emailAddress, password } = this.state;
     return (
-      <div className='row'>
-        <div className='col-xs-12 text-center'>
-          <h1 className='title'>MERNjs</h1>
-        </div>
-        <div className='col-xs-4 col-xs-offset-4'>
-          <form onSubmit={this.handleSubmit.bind(this)}>
-            <div className='form-group'>
-              <label htmlFor='emailAddress'>Email address</label>
-              <input
-                ref='emailAddress'
-                type='email'
-                className='form-control'
-                placeholder='Email Address' />
-            </div>
-            <div className='form-group'>
-              <label htmlFor='password'>Password</label>
-              <input
-                ref='password'
-                type='password'
-                className='form-control'
-                placeholder='Password' />
-            </div>
-            <div className='col-xs-offset-6'>
-              <button
-                type='submit'
-                className='btn btn-block btn-primary'>
-                Sign In
+      <div className="row">
+        <div className="col-lg-4 offset-lg-4">
+          <form onSubmit={this.handleFormSubmit}>
+            <fieldset>
+              <legend className="text-center">Login</legend>
+              <div className="form-group">
+                <label htmlFor="emailAddress">Email Address</label>
+                <input
+                  name="emailAddress"
+                  type="email"
+                  className="form-control"
+                  value={emailAddress}
+                  onChange={this.handleInputChange}
+                  placeholder="Enter email address"
+                />
+              </div>
+              <div className="form-group">
+                <label htmlFor="password">Password</label>
+                <input
+                  name="password"
+                  type="password"
+                  className="form-control"
+                  value={password}
+                  onChange={this.handleInputChange}
+                  placeholder="Enter password"
+                />
+              </div>
+              <button type="submit" className="btn btn-block btn-primary">
+                Submit
               </button>
-            </div>
-            <div className='text-center'>
-              <p><br/>- OR -<br/></p>
-              <Link
-                to='/admin/auth/register'
-                className='btn btn-block btn-default'>
+              <hr />
+              <Link to="/admin/auth/register" className="btn btn-block btn-dark">
                 Register a new account
               </Link>
-            </div>
+            </fieldset>
           </form>
         </div>
       </div>
     );
   }
-
-  handleSubmit(event) {
-    const _this = this;
-    event.preventDefault();
-
-    let emailAddress = ReactDOM.findDOMNode(_this.refs.emailAddress).value.trim();
-    let password = ReactDOM.findDOMNode(_this.refs.password).value.trim();
-
-    /*
-     * Validation rules
-     */
-    if (emailAddress.length < 1) {
-      toastr.warning('Email Address is required.', 'MERNjs');
-      return false;
-    }
-
-    if (password.length < 1) {
-      toastr.warning('Password is required.', 'MERNjs');
-      return false;
-    }
-
-    this.props.loginUser({
-      emailAddress, password,
-    }, function (err, res) {
-      if (err) {
-        console.error('loginUser: ', err);
-        toastr.error(err.message, 'MERNjs');
-      } else {
-        console.log('loginUser: ', res);
-        toastr.success(res.message, 'MERNjs');
-        browserHistory.push('/admin');
-      }
-    });
-  }
 }
+
+LoginComponent.propTypes = {
+  loginUser: PropTypes.func.isRequired,
+  history: PropTypes.object.isRequired,
+};
 
 export default LoginComponent;
