@@ -1,54 +1,64 @@
 import React, { Component } from 'react';
-import { Link } from 'react-router';
+import PropTypes from 'prop-types';
 import moment from 'moment';
-import toastr from 'toastr';
 
 class BlogComponent extends Component {
-  render() {
-    return (
-      <div className='row'>
-        <div className='col-xs-12 text-center'>
-          <div className='table-responsive'>
-            <table className='table table-bordered table-striped table-hover'>
-              <thead>
-                <tr>
-                  <td>SL</td>
-                  <td>Title</td>
-                  <td>Status</td>
-                  <td>Created By</td>
-                  <td>Created At</td>
-                  <td>Updated At</td>
-                </tr>
-              </thead>
-              <tbody>
-                {this.renderAllBlog()}
-              </tbody>
-            </table>
-          </div>
-        </div>
-      </div>
-    );
-  }
-
   renderAllBlog() {
-    if (this.props.allBlog.length) {
-      return this.props.allBlog.map(function (blog, index) {
+    const { blogs } = this.props;
+    if (blogs.length) {
+      return blogs.map((blog, idx) => {
         return (
-          <tr key={blog._id}>
-            <td>{index + 1}</td>
+          <tr key={idx.toString()}>
+            <td>{idx + 1}</td>
             <td>{blog.title}</td>
-            <td>{blog.status}</td>
-            <td>{blog._createdBy.emailAddress}</td>
-            <td>{moment(blog.createdAt).format('MMMM Do YYYY, h:mm:ss a')}</td>
+            <td>{blog.description}</td>
+            <td>{blog._createdBy.fullName}</td>
             <td>{moment(blog.updatedAt).format('MMMM Do YYYY, h:mm:ss a')}</td>
+            <td>
+              {blog.status.toLowerCase() === 'active' ? (
+                <i className="fa fa-eye text-success" />
+              ) : (
+                <i className="fa fa-eye-slash text-danger" />
+              )}
+            </td>
           </tr>
         );
       });
-    } else {
-      return null;
     }
+    return (
+      <tr>
+        <td colSpan="6" className="text-center">
+          <h1 className="msg-h1">Currently, there is no blog!</h1>
+        </td>
+      </tr>
+    );
   }
 
+  render() {
+    return (
+      <div className="row">
+        <table className="table table-hover">
+          <thead>
+            <tr className="table-active">
+              <th scope="col">No</th>
+              <th scope="col">Title</th>
+              <th scope="col">Description</th>
+              <th scope="col">Author</th>
+              <th scope="col">Last Update At</th>
+              <th scope="col">
+                <i className="fa fa-info-circle text-info" />
+              </th>
+            </tr>
+          </thead>
+          <tbody>{this.renderAllBlog()}</tbody>
+        </table>
+      </div>
+    );
+  }
 }
+
+BlogComponent.propTypes = {
+  blogs: PropTypes.array.isRequired,
+};
 
 export default BlogComponent;
