@@ -1,12 +1,16 @@
 import React from 'react';
-import { Route, IndexRoute } from 'react-router';
+import { Switch, Route, useRouteMatch } from 'react-router-dom';
 
-import BlogContainer from '../containers/blog.jsx';
+import BlogContainer from '../containers/blog';
+import NotFoundComponent from '../../core/components/notFound';
 
 export default function () {
+  const match = useRouteMatch();
   return (
-    <Route path='blogs'>
-      <IndexRoute component={BlogContainer} />
-    </Route>
+    <Switch>
+      <Route exact path={match.path} component={BlogContainer} />
+      <Route path={`${match.path}/about`} render={() => <p>blogs about page...</p>} />
+      <Route component={NotFoundComponent} />
+    </Switch>
   );
-};
+}
