@@ -1,60 +1,58 @@
-import React, { Component } from 'react';
-import { Link } from 'react-router';
+import React from 'react';
+import { Link } from 'react-router-dom';
 
-class NavbarComponent extends Component {
-  render() {
-    return (
-      <nav className='navbar navbar-default navbar-fixed-top'>
-        <div className='container-fluid'>
-          <div className='navbar-header'>
-            <Link to='/admin' className='navbar-brand'>
-              MERNjs <small>admin</small>
+const NavbarComponent = () => (
+  <nav className="navbar navbar-expand-lg fixed-top navbar-light bg-light">
+    <div className="container">
+      <Link to="/admin" className="navbar-brand">
+        MERNjs
+      </Link>
+      <button
+        className="navbar-toggler"
+        type="button"
+        data-toggle="collapse"
+        data-target="#navbarColor03"
+        aria-controls="navbarColor03"
+        aria-expanded="false"
+        aria-label="Toggle navigation"
+      >
+        <span className="navbar-toggler-icon" />
+      </button>
+
+      <div className="collapse navbar-collapse" id="navbarColor03">
+        <ul className="navbar-nav mr-auto">
+          <li className="nav-item active">
+            <Link to="/admin/dashboard" className="nav-link">
+              Dashboard
             </Link>
-          </div>
-          <ul className='nav navbar-nav'>
-            <li><Link to='/admin/blogs'>Blog</Link></li>
-            <li><Link to='/admin/users'>User</Link></li>
-          </ul>
-          {localStorage.getItem('email') ?
-            <ul className='nav navbar-nav navbar-right'>
-              <li className='dropdown'>
-                <Link
-                  className='dropdown-toggle'
-                  data-toggle='dropdown'
-                  role='button'
-                  aria-haspopup='true'
-                  aria-expanded='false'>
-                  <i className='fa fa-envelope' aria-hidden='true'></i>
-                  &nbsp;{localStorage.getItem('email')}&nbsp;
-                  <span className='caret'></span>
-                </Link>
-                <ul className='dropdown-menu'>
-                  <li>
-                    <Link to='/admin/profile'>
-                      <i className='fa fa-user-circle-o' aria-hidden='true'></i> Profile
-                    </Link>
-                  </li>
-                  <li role='separator' className='divider'></li>
-                  <li>
-                    <Link to='/admin/auth/logout'>
-                      <i className='fa fa-sign-out' aria-hidden='true'></i> Logout
-                    </Link>
-                  </li>
-                </ul>
-              </li>
-            </ul> :
-            <ul className='nav navbar-nav navbar-right'>
-              <li>
-                <Link to='/admin'>
-                  <i className='fa fa-sign-in' aria-hidden='true'></i> Login
-                </Link>
-              </li>
-            </ul>
-          }
-        </div>
-      </nav>
-    );
-  }
-}
+          </li>
+          <li className="nav-item active">
+            <Link to="/admin/auth" className="nav-link">
+              Auth
+            </Link>
+          </li>
+          <li className="nav-item active">
+            <Link to="/admin/users" className="nav-link">
+              User
+            </Link>
+          </li>
+          <li className="nav-item active">
+            <Link to="/admin/blogs" className="nav-link">
+              Blog
+            </Link>
+          </li>
+        </ul>
+        <ul className="navbar-nav navbar-right">
+          <li className="nav-item active">
+            <Link to="/admin/auth/logout" className="nav-link">
+              <i className="fa fa-sign-out" aria-hidden="true" />
+              &nbsp;Logout
+            </Link>
+          </li>
+        </ul>
+      </div>
+    </div>
+  </nav>
+);
 
 export default NavbarComponent;
