@@ -1,18 +1,18 @@
 import React, { Component } from 'react';
-import toastr from 'toastr';
 
 class DashboardComponent extends Component {
   componentDidMount() {
-    toastr.info('Welcome to MERNjs.', 'Admin | MERNjs');
+    console.log('Welcome to MERNjs!');
   }
 
   render() {
     return (
-      <div className='row'>
-        <div className='col-xs-12 text-center'>
-          <img src='/images/mernjs.png' width='150px' alt='MERNjs' />
-          <h1>Welcome to MERNjs</h1>
-          <p>Admin Dashboard <small> in {process.env.NODE_ENV}</small></p>
+      <div className="row">
+        <div className="col-lg-12 text-center">
+          <img src="/images/logo.png" width="150px" alt="MERNjs" />
+          <br />
+          <h1 className="msg-h1">Welcome to MERNjs!</h1>
+          <p>an open-source full-stack JavaScript solution</p>
         </div>
       </div>
     );
