@@ -1,30 +1,18 @@
-import React, { Component } from 'react';
+import React from 'react';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 
-import DashboardComponent from '../components/Dashboard.jsx';
+import DashboardComponent from '../components/Dashboard';
 
-class DashboardContainer extends Component {
-  render() {
-    return (
-      <DashboardComponent />
-    );
-  }
-}
+const DashboardContainer = () => <DashboardComponent />;
 
 // Get apps store and pass it as props to DashboardContainer
 //  > whenever store changes, the DashboardContainer will automatically re-render
-function mapStateToProps(store) {
-  return {
-  };
-}
+// eslint-disable-next-line no-unused-vars
+const mapStateToProps = (store) => ({});
 
 // Get actions and pass them as props to to DashboardContainer
-function matchDispatchToProps(dispatch) {
-  return bindActionCreators({
-  }, dispatch);
-}
-
+const matchDispatchToProps = (dispatch) => bindActionCreators({}, dispatch);
 // We don't want to return the plain DashboardContainer (component) anymore,
 // we want to return the smart Container
 //  > DashboardContainer is now aware of state and actions
