@@ -1,19 +1,16 @@
 import React from 'react';
-import { Route, IndexRoute } from 'react-router';
-import toastr from 'toastr';
+import { Switch, Route, useRouteMatch } from 'react-router-dom';
 
-import DashboardContainer from '../containers/Dashboard.jsx';
-
-function isAuth(nextState, replaceState) {
-  if (!localStorage.getItem('token') || !localStorage.getItem('email')) {
-    replaceState('/admin/auth');
-  }
-}
+import { AdminRoute } from '../../core/components/privateRoute';
+import DashboardContainer from '../containers/Dashboard';
+import NotFoundComponent from '../../core/components/notFound';
 
 export default function () {
+  const match = useRouteMatch();
   return (
-    <Route path='dashboard'>
-      <IndexRoute component={DashboardContainer} onEnter={isAuth} />
-    </Route>
+    <Switch>
+      <AdminRoute exact path={match.path} component={DashboardContainer} />
+      <Route component={NotFoundComponent} />
+    </Switch>
   );
 }
