@@ -1,8 +1,8 @@
 import { combineReducers } from 'redux';
 
-import { authReducer } from './auth/reducers/auth.js';
-import { allUserReducer } from './users/reducers/allUser.js';
-import { allBlogReducer } from './blogs/reducers/allBlog.js';
+import authReducer from './auth/reducers/auth';
+import allUserReducer from './users/reducers/allUser';
+import allBlogReducer from './blogs/reducers/allBlog';
 
 const reducers = combineReducers({
   auth: authReducer,
