@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
 
-import { allBlogReducer } from './blogs/reducers/allBlog.js';
+import allBlogReducer from './blogs/reducers/allBlog';
 
 const reducers = combineReducers({
   allBlog: allBlogReducer,
