@@ -1,25 +1,27 @@
-import React, { Component } from 'react';
-import { Router, Route, IndexRoute, IndexRedirect, browserHistory } from 'react-router';
+import React from 'react';
+import { Router, Redirect } from 'react-router';
+import { Route, Switch } from 'react-router-dom';
+import { createBrowserHistory } from 'history';
 
-import CoreLayout from './core/layouts/index.jsx';
-import NotFoundComponent from './core/components/notFound.jsx';
+import CoreLayout from './core/layouts/index';
+import NotFoundComponent from './core/components/notFound';
 
-import homeRoute from './home/routes/index.jsx';
-import blogRoute from './blogs/routes/index.jsx';
+import homeRoute from './home/routes/index';
+import blogRoute from './blogs/routes/index';
 
-class Routes extends Component {
-  render() {
-    return (
-      <Router history={browserHistory}>
-        <Route path='/' component={CoreLayout}>
-          <IndexRedirect to='home'/>
-          {homeRoute()}
-          {blogRoute()}
-          <Route path='*' component={NotFoundComponent} />
+const Routes = () => (
+  <Router history={createBrowserHistory()}>
+    <CoreLayout>
+      <Switch>
+        <Route exact path="/">
+          <Redirect to="/home" />
         </Route>
-      </Router>
-    );
-  }
-}
+        <Route path="/home" component={homeRoute} />
+        <Route path="/blogs" component={blogRoute} />
+        <Route component={NotFoundComponent} />
+      </Switch>
+    </CoreLayout>
+  </Router>
+);
 
 export default Routes;
