@@ -2,11 +2,8 @@ import logger from 'redux-logger';
 import thunk from 'redux-thunk';
 import { createStore, applyMiddleware } from 'redux';
 
-import reducers from './reducers.js';
+import reducers from './reducers';
 
-const store = createStore(
-  reducers,
-  applyMiddleware(thunk, logger())
-);
+const store = createStore(reducers, applyMiddleware(thunk, logger));
 
 export default store;
