@@ -1,35 +1,36 @@
-import Blog from '../models/index.js';
+import Blog from '../models/index';
 
 /**
  * get blogs
  */
 export function getAllBlog(req, res, next) {
-  Blog.find({
-    status: 'Active',
-  }).sort({
-    createdAt: -1,
-  }).populate('_createdBy').exec((err, blogs) => {
-    if (err) return next(err);
-    console.log('getAllBlog(blogs): ', blogs);
-    return res.status(200).json({
-      success: true,
-      message: 'Get all blog',
-      data: blogs,
+  console.log('getAllBlog(): ', req.params, req.body, req.auth);
+  Blog.find({})
+    .sort({
+      createdAt: -1,
+    })
+    .populate('_createdBy', '-password')
+    .exec((err, blogs) => {
+      if (err) return next(err);
+
+      return res.status(200).json({
+        success: true,
+        message: 'Get all blog',
+        data: blogs,
+      });
     });
-  });
 }
 
 /**
  * add blog
  */
 export function addBlog(req, res, next) {
-  console.log('addBlog(req.body): ', req.body);
-  let newBlog = new Blog();
+  console.log('addBlog(): ', req.params, req.body, req.auth);
+  const newBlog = new Blog();
   Object.assign(newBlog, req.body, {
-    _createdBy: req.authUser._id,
+    _createdBy: req.auth._id,
   });
-  console.log('addBlog(newBlog): ', newBlog);
-  newBlog.save(function (err, blog) {
+  newBlog.save((err, blog) => {
     if (err) return next(err);
 
     return res.status(201).json({
@@ -41,16 +42,19 @@ export function addBlog(req, res, next) {
 }
 
 /**
- * get blog
+ * get blog by id
  */
 export function getBlog(req, res, next) {
-  Blog.findById(req.params._id).populate('_createdBy').exec((err, blog) => {
-    if (err) return next(err);
+  console.log('getBlog(): ', req.params, req.body, req.auth);
+  Blog.findById(req.params._id)
+    .populate('_createdBy', '-password')
+    .exec((err, blog) => {
+      if (err) return next(err);
 
-    return res.status(200).json({
-      success: true,
-      message: 'Get blog',
-      data: blog,
+      return res.status(200).json({
+        success: true,
+        message: 'Get blog',
+        data: blog,
+      });
     });
-  });
 }
