@@ -1,54 +1,32 @@
-import bcrypt from 'bcrypt-nodejs';
+/* eslint-disable func-names */
+import bcrypt from 'bcryptjs';
 import mongoose, { Schema } from 'mongoose';
 
 // create a userSchema
-export const userSchema = new Schema({
-  fullName: {
-    type: String,
-    required: [true, 'Full Name is required.'],
+export const userSchema = new Schema(
+  {
+    fullName: {
+      type: String,
+      required: [true, 'Full Name is required.'],
+    },
+    emailAddress: {
+      type: String,
+      required: [true, 'Email Address is required.'],
+      unique: true,
+    },
+    password: {
+      type: String,
+      required: [true, 'Password is required.'],
+      // select: false,
+    },
+    status: {
+      type: String,
+      enum: ['Active', 'Inactive'],
+      default: 'Active',
+    },
   },
-  emailAddress: {
-    type: String,
-    required: [true, 'Email Address is required.'],
-    unique: true,
-  },
-  password: {
-    type: String,
-    required: [true, 'Password is required.'],
-  },
-  status: {
-    type: String,
-    enum: ['Active', 'Inactive'],
-    default: 'Active',
-  },
-  role: {
-    type: String,
-    enum: ['Subscriber', 'Administrator'],
-    default: 'Subscriber',
-  },
-  createdAt: Date,
-  updatedAt: Date,
-}, { versionKey: false });
-
-/*
- * userSchema middlewares
- */
-
-// on every save, add the date
-userSchema.pre('save', function (next) {
-  // get the current date
-  var currentDate = new Date();
-
-  // change the updated_at field to current date
-  this.updatedAt = currentDate;
-
-  // if created_at doesn't exist, add to that field
-  if (!this.createdAt) {
-    this.createdAt = currentDate;
-  }
-
-  next();
-});
+  { timestamps: true, versionKey: false },
+);
 
 /*
  * userSchema methods
