@@ -1,15 +1,11 @@
 import express from 'express';
 
-import { verifyToken } from '../../auth/middlewares/verifyToken.js';
-import { getAllUser, addUser, getUser } from '../controllers/index.js';
+import verifyToken from '../../auth/middlewares/verifyToken';
+import { getAllUser, addUser, getUser } from '../controllers/index';
 
 const userRoutes = express.Router();
-
-userRoutes.route('/')
-  .get(getAllUser)
-  .post(verifyToken, addUser);
-
-userRoutes.route('/:_id')
-  .get(getUser);
+userRoutes.use(verifyToken);
+userRoutes.route('/').get(getAllUser).post(addUser);
+userRoutes.route('/:_id').get(getUser);
 
 export default userRoutes;
