@@ -1,88 +1,47 @@
-var path = require('path');
-var config = require('config');
-var webpack = require('webpack');
-var ExtractTextPlugin = require('extract-text-webpack-plugin');
-var CleanWebpackPlugin = require('clean-webpack-plugin');
-var CompressionPlugin = require('compression-webpack-plugin');
+require('dotenv').config();
+const path = require('path');
+const webpack = require('webpack');
+const DotENV = require('dotenv-webpack');
+const { CleanWebpackPlugin } = require('clean-webpack-plugin');
+const TerserJSPlugin = require('terser-webpack-plugin');
+const MiniCssExtractPlugin = require('mini-css-extract-plugin');
+const OptimizeCSSAssetsPlugin = require('optimize-css-assets-webpack-plugin');
+const CompressionPlugin = require('compression-webpack-plugin');
 
-var constants = {};
-if (config.util.getEnv('NODE_ENV') == 'production') {
-  constants = {
-    devtool: 'cheap-module-source-map',
-    output: {
-      filename: '[name].bundle.min.js',
-      sourceMapFilename: '[name].bundle.min.map',
-    },
-    plugins: {
-      filename: '[name].bundle.min.css',
-    },
-  };
-} else {
-  constants = {
-    devtool: 'eval',
-    output: {
-      filename: '[name].bundle.js',
-      sourceMapFilename: '[name].bundle.map',
-    },
-    plugins: {
-      filename: '[name].bundle.css',
-    },
-  };
-}
-
-var plugins = [
-  new CleanWebpackPlugin(['dist'], {
-    root: path.join(__dirname, 'public'),
-    verbose: true,
-  }),
-  new webpack.DefinePlugin({
-    'process.env': {
-      NODE_ENV: JSON.stringify(config.util.getEnv('NODE_ENV')),
-      BASE_URL: JSON.stringify(config.get('BASE_URL')),
-      API_URL: JSON.stringify(config.get('API_URL')),
-    },
-  }),
-  new ExtractTextPlugin(constants.plugins.filename),
+const pluginList = [
+  new DotENV(),
+  new CleanWebpackPlugin(),
   new webpack.ProvidePlugin({
     $: 'jquery',
     jQuery: 'jquery',
   }),
+  new MiniCssExtractPlugin({
+    filename: process.env.NODE_ENV === 'production' ? '[name].bundle.min.css' : '[name].bundle.css',
+    chunkFilename: process.env.NODE_ENV === 'production' ? '[id].bundle.min.css' : '[id].bundle.css',
+    ignoreOrder: false,
+  }),
 ];
 
-if (config.util.getEnv('NODE_ENV') == 'production') {
-  plugins.push(
-    new webpack.optimize.UglifyJsPlugin({
-      compressor: {
-        warnings: false,
-      },
-      output: {
-        comments: false,
-      },
-      sourceMap: true,
-    })
-  );
-  plugins.push(new webpack.optimize.AggressiveMergingPlugin());
-  plugins.push(
-    new CompressionPlugin({
-      asset: '[path].gz[query]',
-      algorithm: 'gzip',
-      test: /\.js$|\.css$|\.html$/,
-      threshold: 10240,
-      minRatio: 0.8,
-    })
-  );
+if (process.env.NODE_ENV === 'production') {
+  pluginList.push(new CompressionPlugin());
 }
 
 module.exports = {
-  devtool: constants.devtool,
+  devtool: process.env.NODE_ENV === 'production' ? '' : 'source-map',
   entry: {
-    client: './client/index.jsx',
-    admin: './admin/index.jsx',
+    client: path.join(__dirname, 'client', 'index.jsx'),
+    admin: path.join(__dirname, 'admin', 'index.jsx'),
   },
   output: {
-    path: path.join(__dirname, './public/dist'),
-    filename: constants.output.filename,
-    sourceMapFilename: constants.output.sourceMapFilename,
+    path: path.join(__dirname, 'public/dist'),
+    filename: process.env.NODE_ENV === 'production' ? '[name].bundle.min.js' : '[name].bundle.js',
+  },
+  mode: process.env.NODE_ENV || 'development',
+  resolve: {
+    extensions: ['.js', '.jsx'],
+  },
+  optimization: {
+    minimizer: [new TerserJSPlugin({}), new OptimizeCSSAssetsPlugin({})],
   },
   module: {
     rules: [
@@ -90,32 +49,29 @@ module.exports = {
         test: /\.(js|jsx)$/,
         exclude: /node_modules/,
         use: ['babel-loader'],
-      }, {
+      },
+      {
         test: /\.(css|scss)$/,
-        use: ExtractTextPlugin.extract({
-          fallback: 'style-loader',
-          use: ['css-loader', 'sass-loader'],
-        }),
-      }, {
-        test: /\.(png|jpg|jpeg|gif)$/,
-        use: {
-          loader: 'url-loader',
-          options: {
-            limit: 100000,
-            name: 'images/[name].[ext]',
+        use: [
+          {
+            loader: MiniCssExtractPlugin.loader,
+            options: {
+              hmr: process.env.NODE_ENV === 'development',
+            },
           },
-        },
-      }, {
+          'css-loader',
+          'sass-loader',
+        ],
+      },
+      {
+        test: /\.(jpg|jpeg|png|gif|mp3)$/,
+        use: ['file-loader'],
+      },
+      {
         test: /\.(svg|eot|woff|woff2|ttf)(\?.*$|$)/,
-        use: {
-          loader: 'url-loader',
-          options: {
-            limit: 100000,
-            name: 'fonts/[name].[ext]',
-          },
-        },
+        use: ['url-loader'],
       },
     ],
   },
-  plugins: plugins,
+  plugins: pluginList,
 };
